requirements.txt
openai>=1.40.0
pdfplumber>=0.11.0
pydantic>=2.7.0
tqdm>=4.66.0
python-dotenv>=1.0.1
# for token counting (with graceful fallback)
tiktoken>=0.7.0
